# DOCUMENTACAO - openApi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rodízio Carros",
    "version": "1.0",
    "description": "Backend de cadastro e consulta de carros permitidos para utilziar de acordo com a os dias de rodízio.",
    "contact": {
      "name": "João Higo Sousa Nunes",
      "email": "higo.sousaa@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/veiculo/{veiculo}/marcas": {
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "description": "",
                  "minItems": 1,
                  "uniqueItems": true,
                  "items": {
                    "type": "object",
                    "required": [
                      "codigo",
                      "nome"
                    ],
                    "properties": {
                      "codigo": {
                        "type": "string",
                        "minLength": 1
                      },
                      "nome": {
                        "type": "string",
                        "minLength": 1
                      }
                    }
                  },
                  "x-examples": {
                    "example-1": [
                      {
                        "codigo": "1",
                        "nome": "Acura"
                      },
                      {
                        "codigo": "2",
                        "nome": "Agrale"
                      },
                      {
                        "codigo": "3",
                        "nome": "Alfa Romeo"
                      },
                      {
                        "codigo": "4",
                        "nome": "AM Gen"
                      },
                      {
                        "codigo": "5",
                        "nome": "Asia Motors"
                      },
                      {
                        "codigo": "189",
                        "nome": "ASTON MARTIN"
                      },
                      {
                        "codigo": "6",
                        "nome": "Audi"
                      },
                      {
                        "codigo": "207",
                        "nome": "Baby"
                      },
                      {
                        "codigo": "7",
                        "nome": "BMW"
                      },
                      {
                        "codigo": "8",
                        "nome": "BRM"
                      },
                      {
                        "codigo": "123",
                        "nome": "Bugre"
                      },
                      {
                        "codigo": "10",
                        "nome": "Cadillac"
                      },
                      {
                        "codigo": "161",
                        "nome": "Caoa Chery/Chery"
                      },
                      {
                        "codigo": "11",
                        "nome": "CBT Jipe"
                      },
                      {
                        "codigo": "136",
                        "nome": "CHANA"
                      },
                      {
                        "codigo": "182",
                        "nome": "CHANGAN"
                      },
                      {
                        "codigo": "12",
                        "nome": "Chrysler"
                      },
                      {
                        "codigo": "13",
                        "nome": "Citroën"
                      },
                      {
                        "codigo": "14",
                        "nome": "Cross Lander"
                      },
                      {
                        "codigo": "15",
                        "nome": "Daewoo"
                      },
                      {
                        "codigo": "16",
                        "nome": "Daihatsu"
                      },
                      {
                        "codigo": "17",
                        "nome": "Dodge"
                      },
                      {
                        "codigo": "147",
                        "nome": "EFFA"
                      },
                      {
                        "codigo": "18",
                        "nome": "Engesa"
                      },
                      {
                        "codigo": "19",
                        "nome": "Envemo"
                      },
                      {
                        "codigo": "20",
                        "nome": "Ferrari"
                      },
                      {
                        "codigo": "21",
                        "nome": "Fiat"
                      },
                      {
                        "codigo": "149",
                        "nome": "Fibravan"
                      },
                      {
                        "codigo": "22",
                        "nome": "Ford"
                      },
                      {
                        "codigo": "190",
                        "nome": "FOTON"
                      },
                      {
                        "codigo": "170",
                        "nome": "Fyber"
                      },
                      {
                        "codigo": "199",
                        "nome": "GEELY"
                      },
                      {
                        "codigo": "23",
                        "nome": "GM - Chevrolet"
                      },
                      {
                        "codigo": "153",
                        "nome": "GREAT WALL"
                      },
                      {
                        "codigo": "24",
                        "nome": "Gurgel"
                      },
                      {
                        "codigo": "152",
                        "nome": "HAFEI"
                      },
                      {
                        "codigo": "214",
                        "nome": "HITECH ELECTRIC"
                      },
                      {
                        "codigo": "25",
                        "nome": "Honda"
                      },
                      {
                        "codigo": "26",
                        "nome": "Hyundai"
                      },
                      {
                        "codigo": "27",
                        "nome": "Isuzu"
                      },
                      {
                        "codigo": "208",
                        "nome": "IVECO"
                      },
                      {
                        "codigo": "177",
                        "nome": "JAC"
                      },
                      {
                        "codigo": "28",
                        "nome": "Jaguar"
                      },
                      {
                        "codigo": "29",
                        "nome": "Jeep"
                      },
                      {
                        "codigo": "154",
                        "nome": "JINBEI"
                      },
                      {
                        "codigo": "30",
                        "nome": "JPX"
                      },
                      {
                        "codigo": "31",
                        "nome": "Kia Motors"
                      },
                      {
                        "codigo": "32",
                        "nome": "Lada"
                      },
                      {
                        "codigo": "171",
                        "nome": "LAMBORGHINI"
                      },
                      {
                        "codigo": "33",
                        "nome": "Land Rover"
                      },
                      {
                        "codigo": "34",
                        "nome": "Lexus"
                      },
                      {
                        "codigo": "168",
                        "nome": "LIFAN"
                      },
                      {
                        "codigo": "127",
                        "nome": "LOBINI"
                      },
                      {
                        "codigo": "35",
                        "nome": "Lotus"
                      },
                      {
                        "codigo": "140",
                        "nome": "Mahindra"
                      },
                      {
                        "codigo": "36",
                        "nome": "Maserati"
                      },
                      {
                        "codigo": "37",
                        "nome": "Matra"
                      },
                      {
                        "codigo": "38",
                        "nome": "Mazda"
                      },
                      {
                        "codigo": "211",
                        "nome": "Mclaren"
                      },
                      {
                        "codigo": "39",
                        "nome": "Mercedes-Benz"
                      },
                      {
                        "codigo": "40",
                        "nome": "Mercury"
                      },
                      {
                        "codigo": "167",
                        "nome": "MG"
                      },
                      {
                        "codigo": "156",
                        "nome": "MINI"
                      },
                      {
                        "codigo": "41",
                        "nome": "Mitsubishi"
                      },
                      {
                        "codigo": "42",
                        "nome": "Miura"
                      },
                      {
                        "codigo": "43",
                        "nome": "Nissan"
                      },
                      {
                        "codigo": "44",
                        "nome": "Peugeot"
                      },
                      {
                        "codigo": "45",
                        "nome": "Plymouth"
                      },
                      {
                        "codigo": "46",
                        "nome": "Pontiac"
                      },
                      {
                        "codigo": "47",
                        "nome": "Porsche"
                      },
                      {
                        "codigo": "185",
                        "nome": "RAM"
                      },
                      {
                        "codigo": "186",
                        "nome": "RELY"
                      },
                      {
                        "codigo": "48",
                        "nome": "Renault"
                      },
                      {
                        "codigo": "195",
                        "nome": "Rolls-Royce"
                      },
                      {
                        "codigo": "49",
                        "nome": "Rover"
                      },
                      {
                        "codigo": "50",
                        "nome": "Saab"
                      },
                      {
                        "codigo": "51",
                        "nome": "Saturn"
                      },
                      {
                        "codigo": "52",
                        "nome": "Seat"
                      },
                      {
                        "codigo": "183",
                        "nome": "SHINERAY"
                      },
                      {
                        "codigo": "157",
                        "nome": "smart"
                      },
                      {
                        "codigo": "125",
                        "nome": "SSANGYONG"
                      },
                      {
                        "codigo": "54",
                        "nome": "Subaru"
                      },
                      {
                        "codigo": "55",
                        "nome": "Suzuki"
                      },
                      {
                        "codigo": "165",
                        "nome": "TAC"
                      },
                      {
                        "codigo": "56",
                        "nome": "Toyota"
                      },
                      {
                        "codigo": "57",
                        "nome": "Troller"
                      },
                      {
                        "codigo": "58",
                        "nome": "Volvo"
                      },
                      {
                        "codigo": "59",
                        "nome": "VW - VolksWagen"
                      },
                      {
                        "codigo": "163",
                        "nome": "Wake"
                      },
                      {
                        "codigo": "120",
                        "nome": "Walk"
                      }
                    ]
                  }
                },
                "examples": {
                  "example-1": {
                    "value": [
                      {
                        "codigo": "1",
                        "nome": "Acura"
                      },
                      {
                        "codigo": "2",
                        "nome": "Agrale"
                      },
                      {
                        "codigo": "3",
                        "nome": "Alfa Romeo"
                      },
                      {
                        "codigo": "4",
                        "nome": "AM Gen"
                      }
                    ]
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string",
                      "minLength": 1
                    },
                    "path": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "timestamp",
                    "status",
                    "error",
                    "path"
                  ],
                  "x-examples": {
                    "example-1": {
                      "timestamp": "2021-06-16T00:16:54.563+00:00",
                      "status": 403,
                      "error": "Forbidden",
                      "path": "/api/veiculo/carros/marcas"
                    }
                  }
                },
                "examples": {
                  "example-1": {
                    "value": {
                      "timestamp": "2021-06-16T00:16:54.563+00:00",
                      "status": 403,
                      "error": "Forbidden",
                      "path": "/api/veiculo/carros/marcas"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-api-veiculo-carros-marcas",
        "security": [
          {
            "Bearer token": []
          }
        ],
        "description": "Lista todas as marcas dos veículos"
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "veiculo",
          "in": "path",
          "required": true,
          "description": "Tipo de veídulo (carros, motos, caminhoes)"
        }
      ]
    },
...
    "/api/veiculo/{veiculo}/marcas/{idmarca}/modelos/{idmodelo}/anos": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "veiculo",
          "in": "path",
          "required": true,
          "description": "Tipo de veídulo (carros, motos, caminhoes)"
        },
        {
          "schema": {
            "type": "number"
          },
          "name": "idmarca",
          "in": "path",
          "required": true,
          "description": "id da marca"
        },
        {
          "schema": {
            "type": "number"
          },
          "name": "idmodelo",
          "in": "path",
          "required": true,
          "description": "id do modelo"
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "description": "",
                  "minItems": 1,
                  "uniqueItems": true,
                  "items": {
                    "type": "object",
                    "required": [
                      "codigo",
                      "nome"
                    ],
                    "properties": {
                      "codigo": {
                        "type": "string",
                        "minLength": 1
                      },
                      "nome": {
                        "type": "string",
                        "minLength": 1
                      }
                    }
                  },
                  "x-examples": {
                    "example-1": [
                      {
                        "codigo": "32000-3",
                        "nome": "32000 Diesel"
                      },
                      {
                        "codigo": "2020-3",
                        "nome": "2020 Diesel"
                      },
                      {
                        "codigo": "2019-3",
                        "nome": "2019 Diesel"
                      },
                      {
                        "codigo": "2018-3",
                        "nome": "2018 Diesel"
                      },
                      {
                        "codigo": "2017-3",
                        "nome": "2017 Diesel"
                      },
                      {
                        "codigo": "2016-3",
                        "nome": "2016 Diesel"
                      },
                      {
                        "codigo": "2015-3",
                        "nome": "2015 Diesel"
                      },
                      {
                        "codigo": "2014-3",
                        "nome": "2014 Diesel"
                      },
                      {
                        "codigo": "2013-3",
                        "nome": "2013 Diesel"
                      },
                      {
                        "codigo": "2012-3",
                        "nome": "2012 Diesel"
                      }
                    ]
                  }
                },
                "examples": {
                  "example-1": {
                    "value": [
                      {
                        "codigo": "32000-3",
                        "nome": "32000 Diesel"
                      },
                      {
                        "codigo": "2020-3",
                        "nome": "2020 Diesel"
                      },
                      {
                        "codigo": "2019-3",
                        "nome": "2019 Diesel"
                      }
                    ]
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string",
                      "minLength": 1
                    },
                    "path": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "timestamp",
                    "status",
                    "error",
                    "path"
                  ],
                  "x-examples": {
                    "example-1": {
                      "timestamp": "2021-06-16T02:49:07.872+00:00",
                      "status": 403,
                      "error": "Forbidden",
                      "path": "/api/veiculo/carros/marcas/59/modelos/5940/anos"
                    }
                  }
                },
                "examples": {
                  "example-1": {
                    "value": {
                      "timestamp": "2021-06-16T02:49:07.872+00:00",
                      "status": 403,
                      "error": "Forbidden",
                      "path": "/api/veiculo/carros/marcas/59/modelos/5940/anos"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-api-veiculo-veiculo-marcas-idmarca-modelos-idmodelo-anos",
        "security": [
          {
            "Bearer token": []
          }
        ]
      }
    },
    "/api/veiculo/{veiculo}/marcas/{idmarca}/modelos/{idmodelo}/anos/{ano}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "veiculo",
          "in": "path",
          "required": true,
          "description": "Tipo de veídulo (carros, motos, caminhoes)"
        },
        {
          "schema": {
            "type": "number"
          },
          "name": "idmarca",
          "in": "path",
          "required": true,
          "description": "id da marca"
        },
        {
          "schema": {
            "type": "number"
          },
          "name": "idmodelo",
          "in": "path",
          "required": true,
          "description": "id do modelo"
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "ano",
          "in": "path",
          "required": true,
          "description": "ano do modelo"
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "type": "object",
                  "properties": {
                    "valor": {
                      "type": "string",
                      "minLength": 1
                    },
                    "marca": {
                      "type": "string",
                      "minLength": 1
                    },
                    "modelo": {
                      "type": "string",
                      "minLength": 1
                    },
                    "combustivel": {
                      "type": "string",
                      "minLength": 1
                    },
                    "codigoFipe": {
                      "type": "string",
                      "minLength": 1
                    },
                    "mesReferencia": {
                      "type": "string",
                      "minLength": 1
                    },
                    "anoModelo": {
                      "type": "number"
                    },
                    "tipoVeiculo": {
                      "type": "number"
                    },
                    "siglaCombustivel": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "valor",
                    "marca",
                    "modelo",
                    "combustivel",
                    "codigoFipe",
                    "mesReferencia",
                    "anoModelo",
                    "tipoVeiculo",
                    "siglaCombustivel"
                  ],
                  "x-examples": {
                    "example-1": {
                      "valor": "R$ 116.925,00",
                      "marca": "VW - VolksWagen",
                      "modelo": "AMAROK High.CD 2.0 16V TDI 4x4 Dies. Aut",
                      "combustivel": "Diesel",
                      "codigoFipe": "005340-6",
                      "mesReferencia": "junho de 2021 ",
                      "anoModelo": 2015,
                      "tipoVeiculo": 1,
                      "siglaCombustivel": "D"
                    }
                  }
                },
                "examples": {
                  "example-1": {
                    "value": {
                      "valor": "R$ 116.925,00",
                      "marca": "VW - VolksWagen",
                      "modelo": "AMAROK High.CD 2.0 16V TDI 4x4 Dies. Aut",
                      "combustivel": "Diesel",
                      "codigoFipe": "005340-6",
                      "mesReferencia": "junho de 2021 ",
                      "anoModelo": 2015,
                      "tipoVeiculo": 1,
                      "siglaCombustivel": "D"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string",
                      "minLength": 1
                    },
                    "path": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "timestamp",
                    "status",
                    "error",
                    "path"
                  ],
                  "x-examples": {
                    "example-1": {
                      "timestamp": "2021-06-16T02:51:22.568+00:00",
                      "status": 403,
                      "error": "Forbidden",
                      "path": "/api/veiculo/carros/marcas/59/modelos/5940/anos/2015-3"
                    }
                  }
                },
                "examples": {
                  "example-1": {
                    "value": {
                      "timestamp": "2021-06-16T02:51:22.568+00:00",
                      "status": 403,
                      "error": "Forbidden",
                      "path": "/api/veiculo/carros/marcas/59/modelos/5940/anos/2015-3"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-api-veiculo-veiculo-marcas-idmarca-modelos-idmodelo-anos-ano",
        "security": [
          {
            "Bearer token": []
          }
        ]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "Bearer token": {
        "name": "token",
        "type": "apiKey",
        "in": "header",
        "description": "Authorization\n"
      }
    }
  },
  "security": [
    {
      "Bearer ": []
    }
  ]
}